backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
sqlalchemy
alembic
psycopg2-binary
redis
celery[redis]
python-dotenv
PyGithub
openai
requests
gunicorn
